Notes and code to test daemon(3) and pbpaste
This is aimed at trying to come up with a working patch to get
pasteboard access inside tmux.

There are lots of combinations present in the code.

I have just been commenting out what I did not want at the time.

I need to test some of this still on 10.4 (at least daemon(3), since
the other vproc stuff may not be available there).