wrapper: 10.5 does not have RTLD_SELF
Mac OS X 10.5 only has RTLD_NEXT and RTLD_DEFAULT. So, switch to
NEXT for now.

For what it is worth, I actually noticed this problem on 10.4.
I checked the manpages online and the 10.5 manpage only mentions
SELF and DEFAULT (same as 10.4). The 10.4 status may or may not
matter in the long run since I still do not have a way of
"reattaching" on 10.4.