wrapper: learn the "-l" option
It makes the exec'd program have an argv[0] that starts with a '-'.
This is usually interpreted by shells as an indication that they
should startup as a "login shell".

We can use this with tmux to have it create its shells in the user
bootstrap namespace (even if it is not there itself).

    tmux set-option -g \
      default-command '/path/to/wrapper -l /path/to/shell'