TODO: -detach +runtimeconf
_vprocmgr_detach_from_console does not do the same kind of work as
_vprocmgr_move_subset_to_user.

From what I can tell, it only prevents its caller (and its future(?)
children) from losing access to the per-user bootstrap namespace
when the process that started their (bootstrap-(?), certainly not
Unix-) session exits.

It will not do anything useful if used with the system daemon(3),
because that daemon() will not only "detach from the console", but
goes even further by moving to the root bootstrap namespace (which
prevents access to the per-user bootstrap namespace). It only seems
useful in places where the system daemon(3) is not used (screen(1)
seems to do without daemon(3), which is why it is useful there).

It works if you use it with your own custom daemon() that does not
forcibly move the resulting process to the root bootstrap namespace,
but it will not reattach a process to the per-user bootstrap
namespace.

So, if we use the system daemon(3) we need to reattach to the
per-user bootstrap namespace with code similar to the 10.5 screen
patch (_vprocmgr_move_subset_to_user). If we use a custom daemon(),
then we use code similar to the 10.6 screen patch
(_vprocmgr_detach_from_console).

Ultimately, we need a private call if we want to ensure access to
the pasteboard after the session-starter has exited.

----

Start plans for revamping 'test', so that we can customize what it
does at runtime (to try out different variations and orders of
calling the various functions without editing and recompiling).