wrapper: proper free if exec and die fail
This is unlikely, but we might as well fix it so that we are not
freeing things we did not allocate.

A development version of 1d628242 (wrapper: learn the "-l" option,
2011-03-18) worked something like this, but it was lost to reworking
before the final commit.

valgrind will point out of the problem if we arrange for die_errno
to not actually exit.