wrapper: provisional 10.7 "Lion" support
Update the wrapper code to not issue a warning about an "unsupported
new OS" when run on Lion.

This change is provisional because I have not tested it on a 10.7
machine. The evidence seems to indicate that the "10.6 method" will
also work on 10.7.

Specifically, The signature of _vprocmgr_move_subset_to_user matches
the one from 10.6 (see libvproc.c or vproc_priv.h in the launchd
source code).

    http://www.opensource.apple.com/release/mac-os-x-107/

Wikipedia's Darwin article says 10.7 is Darwin 11.0.0, so I assume
that is what uname reports for ustname.release.

    http://en.wikipedia.org/wiki/Darwin_(operating_system)