wrapper: 10.8 "Mountain Lion" support
There has been some promising feedback regarding Mountain Lion.

    https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard/commit/f4a2901bb2b4c29a36e3ec0601219d384c5d24e3#all_commit_comments

Update the wrapper code to not issue a warning about an "unsupported
new OS" when run on Mountain Lion.

There may be further changes once Apple publishes the 10.8 launchd
sources (e.g. if the signature of the relevant function is changed
as it was between 10.5 and 10.6).