wrapper: preliminary 10.8 "Lion" support
Update the wrapper code to not issue a warning about an "unsupported
new OS" when run on Mountain Lion.

This may not even be appropriate. I have not used Mountain Lion or
heard from anyone that has.

Assuming 10.8 works like 10.6-10.7, this should be enough to squelch
the warning. This commit will be rewritten (at least to edit this
commit message) once some evidence comes in as to whether this is
appropriate (i.e. probably after a public release of 10.8 and the
release of the open source bits -- especially the launchd source
where the relevant function is defined).