wrapper: handle malloc() errors for -l processing
Handle malloc() errors that might come up while doing "login"
processing (massaging the new command's argv[0] when the "-l" option
is used).

Skip the string/argument manipulation and just use the strings that
were given in the wrapper's argv. The new program will not find that
its argv[0] starts with a dash, but we will at least give it
a chance to try to run (although it may also face serious problems).