wrapper: 10.9 "Mavericks" support
There has been some promising feedback regarding Mavericks.

    https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard/pull/20#issuecomment-19492844

Update the wrapper code to not issue a warning about an "unsupported
new OS" when run on Mavericks. The existing "10.6 variation" still
seems to be both necessary and sufficient.

There may be further changes once Apple publishes the 10.9 launchd
sources (e.g. if the signature of the relevant function is changed
as it was between 10.5 and 10.6).